requirements.txt
streamlit
opencv-python
requests
google-genai
dotenv